requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==2.2.10
django-crispy-forms==1.10.0
django-environ==0.4.5
h11==0.12.0
hiredis==1.1.0
httptools==0.1.1
hyperlink==21.0.0
idna==3.1
incremental==17.5.0
msgpack==1.0.2
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
pyOpenSSL==20.0.1
python-dotenv==0.15.0
pytz==2020.5
PyYAML==5.3.1
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.12.1
uvicorn==0.13.3
uvloop==0.14.0
watchgod==0.6
websockets==8.1
zope.interface==5.2.0
